ai-video-pipeline/requirements.txt
google-generativeai==0.3.2
edge-tts==6.1.0
python-dotenv==1.0.0
requests==2.31.0
moviepy==1.0.3
pillow==10.0.0
numpy==1.26.4
imageio-ffmpeg==0.5.1
google-api-python-client==2.164.0
google-auth-oauthlib==1.2.1
google-auth-httplib2==0.2.0
